Changed Header.js spacing, deleted custom css in App.css, Removed extra spaces in App.jsx
